ecere/gfx/3D/models/e3d: Reorganized textures saving options
e3dWrite.ec: writeTexture: Use value of E3DOptions::texturesQuery to decide wether or not to output external textures.

It is useful for for GeoPackageExport, since  textures are exported separately.
E3DOptions::texturesQuery is copied onto E3DWriteContext.texturesPrefix and:

  if texturesPrefix is ".", the texture name is written in the model, without
  prefix, resolution and extension;

  if texturesPrefix is any other string, it is used as path prefix to the model
  name, that includes resolution and extension.

  if texturesPrefix is null, "textures" is used as path prefix to the model
  name, that includes resolution and extension.

Note that the latter is currently disabled, since externalTextures is set to false.

TODO: Take precautions not to create a "File:" folder if operating on
TempFiles, for now it should not be a problem since we pass "." or null only
and in the latter case writing does not happen.

E3dOptions, E3DContext, E3DWriteContext: more descriptive name for texturesQuery.

The texturesQuery option is really the path of the directory where textures
should be written or searched for, so texturesPath is a better name.

e3dDefs.ec: Move here E3D(Options|(Write)?Context) and add enumerations for new options.